requirements.txt
langgraph== 0.2.39
langchain== 0.3.4
langchain-openai== 0.2.3
python-dotenv== 1.0.1
pandas== 2.2.3
langchain-community== 0.3.3
psycopg2==2.9.10